Add filter.py with composable filters concept
It occurred to me today that filtering would be really nice if I could
simply compose filters together, and then run them, especially for that
fusion (i.e. do all the work on one thing before moving on, no
duplicating of lists for every filtration). This would be cake in
Haskell, but it took a little bit of head-scratching in Python, and a
mix of OO and FP ideas, because types in Python are basically classes.