Switch filter.ast from string to 'tree' output
The previous/original output was a multiline string, with indentation to
indicate the parent/child relations, and with the intention that the
string would be printed out as a visual representation of the AST. what
I really want is an AST I can easily explore, though, and make decisions
about later, so now it's just easily-walkable pairs and lists. The first
item in any given pair is the type of that node, like "PRED" or "AND",
while the second item is a list of the children of the node.