Change NOT operator from - to ~
In talking over the operators with Ben, he didn't like the use of
__neg__, IIRC, because I'd selected & and |, both bitwise operators in
Python, for the AND and OR operations, and ~ is the bitwise NOT
operation, and it made more sense to go with that. I was not married to
the - symbol, and was happy to bring things in line in this way.