Allow boolean ops only through boolean operators
I have an earlier commit wherein I added op, and the left and right
operand slots in as args to Pred's init, thinking it was better to be
explicit about what could go into a pred, but Ben managed to argue the
point with me that, as an interface to the user, it was better not to
allow creation of boolean operations in the init, but restrict them
instead to the boolean operators provided. Ultimately, I agreed.