Add ability to stringify a Pred
This is a rewriting of history. Ben liked my previously implemented idea
for passing in a name to each Pred on init, and then using it when
stringifying, so we rather quickly recreated that, and added tests
around it, before finishing up the tests here, for the AND and OR cases.
He didn't like the pretty-printing I was initially doing in the ast
method (and neither did I; I removed it a while ago), wanting that to be
its own pretty printing function (totally agree). He didn't mind that my
ast function was a barely-there rewriting of the ast already existing in
the Pred instances, which point to each other through the pred, left,
and right properties, and which the ast method easily walks through
while generating its ast result. I think he didn't quite see the point
of it, though, and in retrospect, I think I'm less interested in it as
well. It was always suspect, and a bit on the chopping block for me. The
things we did in our pair session today raised the axe a bit more. So,
now, stringifying just creates a kind of s-expression-like string.