Add name slot to Pred
Adding a name to a pred makes stringification much nicer, as you get
actual names being ANDed and ORed and NOTed, rather than the str of a
function, which is something like "<function <lambda> at..." Ben had the
good idea of storing the name given, if any, or the string of the pred
itself, it not, which also meant we didn't need to str the pred in str,
but just always use the name instead. It also lead to named and unnamed
versions of our 2 tests on PRED and NOT instances. We skipped out on the
AND and OR instances, because 1) we were a bit all over the place, and
2) it felt like a chore to test them without names.