Parenthesize based on operator precedence analysis
This was definitely a Ben trick. I brought up the problem with
parentheses, which he'd already noticed, and I mentioned the issues I'd
had with that in the past. My first solution, years back, for similar
DSLs in Haskell, was to just parenthesize everything, which was Ben's
first reaction while spiking. I mentioned that it gets absurd when you
do that, and said that in the more recent past, I'd had the idea to step
back, and pattern match on a larger expression. He made the jump from
there, possibly based on past experience, that we could simply do the
grouping based on operator precedence. While I understood, it felt a bit
mentally overwhelming to me. It was just too new and foreign to think of
how the precedence aligned with the need to group in this particular
way, so I let him lead, and now I get it, though I wouldn't say it's at
all comfortable yet.