Begin accessor.py with Accessor class
It occurred to me that a common use for me of Pred instances would be to
check some aspect of some bit of data I looked up in some structure,
like the value in a dictionary. It also occurred to me that if I knew
the type of the thing coming back, I could make particular objects that
did that lookup for me, but then imbued the result with the ability to
make Pred instances. So, if name were a kind of accessor that looked up
the name field in a dict (in my code), then it would be nice if I could
immediately get a pred from that, via, e.g., name.startsWith("The_"), or
name.is("Bob"), or name.contains("ab"), or name.matches("^\d\d-.*"), or
name.search("'(\d\d)'").groups(0).is("23") (maybe?).