Test true/false in simplify and name correctly
I had named them True and False, but those are Python's standard boolean
values, which is not what these are, and so it gets confusing. I'm now
using true and false, but part of me wants to call them alwaysTrue and
alwaysFalse, which I once used in a tech talk, before switching to what
I'd probably really like to call them: const(True) and const(False).