Add >> in as new Pred SEQ op
This is currently just an aping of the & (AND) operator. The choice of
>> (Python's right-shift operator, under the __rshift__ method) follows
from Haskell's "and then" operator (>>), which sequences operations,
albeit, forgetting the results of the first. The better choice would
have been >>=, Haskell's "bind" operator, but in Python, this operator
is the in-place version of >>, which means that whatever I return from
its __irshift__ method is automatically assigned over whatever is in the
left slot of the operation, and then None is returned, which means you
can't even do a >>= b >>= c, because a >>= b is evaluated, the returned
value is assigned to a, None gets returned, and then None >>= c raises.