Add Pred.validate, accepting, using a fix function
So far, I'm only handling the "PRED" case. It has once again become
obvious to me how strange having a fix function is, because it relies on
the fix being side-effecting. Pred is a mostly immutable concept; trying
to fix a value with the passed-in function, when the Pred comes up
False, can only work if you're making a change out in the real world. If
you're testing that the input, a number, is equal to 4, and your fix is
to set it equal to 4, you'll just be creating a new, local value in your
function. Numbers (and strings) are immutable, so you can't assign to a
variable holding on, and think you've changed it. You've just pointed a
reference at a new location in memory. This feature is being added to
allow for fixing things in Autodesk Maya, which is getting very specific
for a library ostensibly about predicates, but it's useful enough, and
very useful to me, currently, and fully optional, so...