Handle AND in Pred.validate
This required a huge amount of testing to get right (now at 100 tests!).
I spun my wheels forever, trying to figure out why the right side
expression in the AND was not seeing the mutated value in the input, but
it turned out that using gte(0) & lte(10) wasn't accounting for the fact
that the value I was testing was in a dict. I had forgotten, and yet,
because of the lack of types, Python was happily comparing a dictionary
with a single key, value pair in it with an integer, and showing me
values that matched up with the failing values before modifying the
value in the dict, greatly confusing the issue ::rolleyes::. It made me
with I could use my accessors idea, but I don't want to pull in that
module just for that wish. I like to test things in isolation.