For posterity: fix functions at the AND/OR level
This commit is just for posterity; I'm removing the work I've done to
add fix function handling to the AND and OR level Preds.

I've decided, after too much strangeness, and competition, not just
between the left and right sides of AND and OR Preds, but also between
them and their parent AND or OR, that this isn't worth it. I think it's
fine to note in the documentation that fix functions can only exist at
the PRED level (and maybe NOT?), and not in an AND or OR. From an impl
standpoint, it's easy enough to just not look for, nor use a fix method
found on an AND or OR predicate.