Remove AND/OR validate fix handling, cleanup tests
I've decided there should be no fix handling at the combinatorial levels
(AND/OR), as it causes too many weird cases. I also filled in the many
cases of existing, failing, and passing tests, and matched up that list
across AND and OR, by rearranging tests, and homogenizing their naming.