Add SEQ handling to validate
This turned out simpler than I'd feared it would, especially with the
decision to not allow fix functions at the AND and OR (and now SEQ)
levels. Some of the validate tests from AND couldn't apply to SEQ,
because there's no time when the left fails entirely, and the right
still does anything. Speaking of, when the left fails, the right still
returns a dict of data, but with only the pred, op, left, and right
fields, never a result or status field, because, when the left fails,
the right is never evaluated.

In terms of reporting, I wish I had sum types, but, instead, I have the
presence or absence of keys, and their values in the nest of dicts to
define things. For example, if there isn't a result, it means (I hope)
that you're down the right side of a failed SEQ. If there is a status
key, that means there was a failure, and also a fix function; it's value
determines whether or not the fix function worked.

To get the unsolved right sides of left-failing SEQs, I added a noSolve
arg to validate, defaulting to False. If you pass True, it blocks any
further solving, but still fills out the entire tree. A report can show
that the node right above this unsolved subtree failed, and therefore,
the subtree, depending on that, could not even be tried.

This completes the work on validate, for now.