Rename "pred" to "ref" in validate results
I was using "pred" to refer to the predicate instance in the dict that
gets returned from validate, but then I got to not, which doesn't have a
left and right child, but just a pred child. The dict result for some of
the combinators has left and right entries, but now I've come to NOT,
and it just has a pred entry. I'm returning left and right in the dicts
for AND, OR, and SEQ, but that leaves me with nowhere to return the pred
result, except pred, which is taken, so I moved all the self-referencing
pred keys to ref keys, so in this one place, I can use pred, so NOT can
refer to its one child. I'm also now testing that lower level in the NOT
tests.