Take, check against type constraint pair in init
For posterity (and later documentation), the typeCon arg to Pred's init
is a pair of either a Python type, or a list thereof, and a predicate on
the eventually tested value. The reason I accept a list of types is that
it's possible that a predicate could work on a set of values across more
than one type, the most obvious being the union of int and float. Also,
the reason I require 2 things (the type or type list and a predicate on
the eventual value) is that init'ing, and also combining Preds together
will requirea checking the type (or list thereof) for sameness, whereas
the eventual call will need to pass the value into the predicate to
allow for the myriad ways you may want to test for the appropriate type:

    type(v) == str
    type(v) is str
    isinstance(v, basestr): allows str, unicode, and derived classes
    type(v) in [int, float]: specifically those two
    isinstance(v, int) or isinstance(v, float): allows derived classes
    isinstance(v, numbers.Number): like basestr, for all number types
    etc...