Store type constraint type lists in sorted order
This will make it easier to compare them, without having to turn them
into a set, or even having to ask if they're a list or not. It does make
me imagine edge cases where things should be compatible, but aren't,
like if one expects a float, and the other, something derived therefrom.

I didn't test this, and I can't imagine it helping, but I also hope that
sorting types in Python is deterministic, at least within a session.