Fix noSolve code
For shame! I found some untested, and completely unexercised code paths
hidden at the top of the validate method. The code in there had 2 lines
with glaring errors, too, namely, that "left" and "right" keys were
unquoted, which would be a runtime error (and the new tests forced it to
be so), and that the NOT branch was returning a "left" - also without
quotes (another would-be error) - instead of a "pred". All fixed now,
with tests.

UPDATE: this is an interactively rebased-in note, added in to say that I
found this comment, tucked away in some stashed changes, in the noSolve
section of the code:

    TODO: test noSolve=True (currently untested)

Looks like I hadn't completely lost my mind.