Fix naming of accessor-created predicates
This change combines the accessor name with automatically generated name
of the predicate from each pred-creating method. Formerly, if you had a
name accessor with the name "name", and you called its startswith method
to create a pred, the name of that pred would be 'startswith("...")'.
Now it will be 'name.startswith("...")'.